+++ nussl/__init__.py
@@ -5,15 +5,24 @@ except Exception:
     vamp_imported = False
 
 # Current nussl version
-__version__ = '1.0.1'
+__version__ = '1.1.0'
 
 class ImportErrorClass(object):
     def __init__(self, lib, **kwargs):
-        raise ImportError(f'Cannot import {type(self).__name__} because {lib} is not installed')
+        raise ImportError(
+            f'Cannot import {type(self).__name__} because {lib} is not installed')
 
+try:
+    import musdb
+except RuntimeError:
+    warnings.warn(
+        "Importing musdb failed, likely because ffmpeg or ffprobe are not installed!.")
+    musdb = None
 
 from .core import AudioSignal, STFTParams
 from .core import utils, efz_utils, play_utils, constants, mixing
+import warnings
+
 
 from . import core
 from . import evaluation