+++ nussl/ml/train/loss.py
@@ -17,21 +17,34 @@ class SISDRLoss(nn.Module):
     Computes the Scale-Invariant Source-to-Distortion Ratio between a batch
     of estimated and reference audio signals. Used in end-to-end networks.
     This is essentially a batch PyTorch version of the function 
-    ``nussl.evaluation.bss_eval.scale_bss_eval``.
+    ``nussl.evaluation.bss_eval.scale_bss_eval`` and can be used to compute
+    SI-SDR or SNR.
     
     Args:
         scaling (bool, optional): Whether to use scale-invariant (True) or
           signal-to-noise ratio (False). Defaults to True.
+        return_scaling (bool, optional): Whether to only return the scaling
+          factor that the estimate gets scaled by relative to the reference. 
+          This is just for monitoring this value during training, don't actually
+          train with it! Defaults to False.
+        reduction (str, optional): How to reduce across the batch (either 'mean', 
+          'sum', or none). Defaults to 'mean'.
+        zero_mean (bool, optional): Zero mean the references and estimates before
+          computing the loss. Defaults to True.
     """
     DEFAULT_KEYS = {'audio': 'estimates', 'source_audio': 'references'}
 
-    def __init__(self, scaling=True, reduction='mean', zero_mean=True):
+    def __init__(self, scaling=True, return_scaling=False, reduction='mean', 
+                 zero_mean=True):
         self.scaling = scaling
         self.reduction = reduction
         self.zero_mean = zero_mean
+        self.return_scaling = return_scaling
         super().__init__()
 
     def forward(self, estimates, references):
+        eps = 1e-10
+        # num_batch, num_samples, num_sources
         _shape = references.shape
         references = references.view(-1, _shape[-2], _shape[-1])
         estimates = estimates.view(-1, _shape[-2], _shape[-1])
@@ -46,13 +59,8 @@ class SISDRLoss(nn.Module):
         _references = references - mean_reference
         _estimates = estimates - mean_estimate
         
-        references_projection = _references.transpose(2, 1) @ _references
-
-        references_projection = torch.diagonal(
-            references_projection, dim1=-2, dim2=-1) + 1e-10
-    
-        references_on_estimates = torch.diagonal(
-            references.transpose(2, 1) @ _estimates, dim1=-2, dim2=-1) + 1e-10
+        references_projection = (_references ** 2).sum(dim=-2) + eps
+        references_on_estimates = (_estimates * _references).sum(dim=-2) + eps
 
         scale = (
             (references_on_estimates / references_projection).unsqueeze(1)
@@ -70,6 +78,8 @@ class SISDRLoss(nn.Module):
             sdr = sdr.mean()
         elif self.reduction == 'sum':
             sdr = sdr.sum()
+        if self.return_scaling:
+            return scale
         # go negative so it's a loss
         return -sdr