+++ tests/evaluation/test_evaluation.py
@@ -257,7 +257,7 @@ def test_scale_bss_eval(estimated_and_true_sources):
     mixture = sum(true_sources).audio_data.T
     mixture -= mixture.mean(axis=0)
 
-    tSISDR, tSISIR, tSISAR, tSDSDR, tSNR, tSRR, tSDRi, tSDSDRi, tSNRi = (
+    tSISDR, tSISIR, tSISAR, tSDSDR, tSNR, tSRR, tSDRi, tSDSDRi, tSNRi, _, _, _ = (
         nussl.evaluation.scale_bss_eval(
             _references, _estimates[..., 0], mixture[..., 0], 0,
         )
@@ -269,7 +269,7 @@ def test_scale_bss_eval(estimated_and_true_sources):
     _references = references[:, 0, :]
     _estimates = estimates[:, 0, :]
 
-    rSISDR, rSISIR, rSISAR, rSDSDR, rSNR, rSRR, tSDRi, tSDSDRi, tSNRi = (
+    rSISDR, rSISIR, rSISAR, rSDSDR, rSNR, rSRR, tSDRi, tSDSDRi, tSNRi, _, _, _ = (
         nussl.evaluation.scale_bss_eval(
             _references, _estimates[..., 0], mixture[..., 0], 0,
         )